rename project pdal -> points
