+++ points/src/main/scala/geotrellis/spark/points/json/package.scala
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2016 Azavea
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package geotrellis.spark.points
+
+import java.io.File
+
+import spray.json._
+import spray.json.DefaultJsonProtocol._
+
+package object json extends MetadataFormat {
+  def fileToPipelineJson(localPath: File): JsObject =
+    JsObject("pipeline" -> JsArray(JsObject("filename" -> localPath.getAbsolutePath.toJson)))
+}
